MSVC: cast the result, not the constant
C4310, "cast truncates constant value". `(u16)~MASK` complements a small
constant to a 32-bit one and then casts it down, so the thing being truncated is
a constant and MSVC says so. Every one of these is deliberate -- the point is to
clear some bits -- but the compiler cannot tell deliberate from accidental when
what it sees is a literal that does not fit.

Cast the whole expression instead: `x = (u16)(x & ~(unsigned)MASK)`. Same value,
and there is no longer a constant in the cast for the warning to land on. Ten
sites, found by grep rather than by CI, because /WX stops MSVC at the first one
and each round trip through Actions costs five minutes to learn one name.

cmd_ai.c keeps its constant -- it is the immediate the check expects to find in
the executable -- so that one is masked to fit rather than restructured.

GCC and Clang still clean with -Werror, 35 tests passing on both.

+++ src/formats/surface.h
@@ -278,8 +278,8 @@ Q2PSX_INLINE bool q2_scene_flags_deferred(u16 flags)
  * touched. Transcribed from 0x8002E7CC-0x8002E7D8. */
 Q2PSX_INLINE u16 q2_scene_flags_set_wibble(u16 flags, u32 value)
 {
-    return (u16)((flags & (u16)~Q2_SCENE_FLAG_WIBBLE_MASK)
-                 | (u16)((value & 0xFu) << Q2_SCENE_FLAG_WIBBLE_SHIFT));
+    return (u16)((flags & ~(unsigned)Q2_SCENE_FLAG_WIBBLE_MASK)
+                 | ((value & 0xFu) << Q2_SCENE_FLAG_WIBBLE_SHIFT));
 }
 
 /* ------------------------------------------------------------------------- */
